Add community tools section (#80)
Add a new page for Community Tools on the PayID docs website.

Context of Change:

I came across this tool today, and was really impressed by its quality. A lot of project websites list high-quality community tools on their documentation website, and I think it's reasonable we do the same.