Update metrics documentation (#89)
Update our metrics documentation now that metrics are opt-out rather than opt-in.

Also refactored things a little bit, and updated the curl command to work.

Context of Change
We want PayID server metrics collection to be opt-out rather than opt-in. Most people will go with whatever is easiest, and we want the easiest path to be reporting metrics.

The PayID server code has already been updated, so this updates documentation to reflect that.